+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/Core/Arbitrary.hs
@@ -38,6 +38,14 @@ module Test.Cardano.Ledger.Core.Arbitrary (
 ) where
 
 import qualified Cardano.Chain.Common as Byron
+import Cardano.Crypto.DSIGN (
+  BLS12381MinSigDSIGN,
+  DSIGNAggregatable (PossessionProofDSIGN, createPossessionProofDSIGN),
+  DSIGNAlgorithm (genKeyDSIGNWithContext),
+  seedSizeDSIGN,
+ )
+import Cardano.Crypto.DSIGN.BLS12381.Internal (minSigPoPDST)
+import Test.Crypto.Util (arbitrarySeedOfSize)
 import Cardano.Crypto.Hash.Class
 import Cardano.Ledger.Address
 import Cardano.Ledger.BaseTypes (
@@ -464,6 +472,7 @@ instance Arbitrary StakePoolParams where
       <*> arbitrary
       <*> arbitrary
       <*> arbitrary
+      <*> pure BaseTypes.SNothing -- sppLeiosKey: only encoded in PV12+
 
 instance Arbitrary StakePoolState where
   arbitrary =
@@ -481,6 +490,23 @@ instance Arbitrary StakePoolState where
 
 instance Arbitrary PoolMetadata where
   arbitrary = PoolMetadata <$> arbitrary <*> arbitrary
+
+instance Arbitrary LeiosKey where
+  arbitrary = LeiosKey <$> arbitrary <*> arbitrary
+
+instance Arbitrary LeiosPubKey where
+  arbitrary = LeiosPubKey <$> arbitrary
+
+instance Arbitrary LeiosPossessionProof where
+  arbitrary = LeiosPossessionProof <$> genLeiosPossessionProof
+
+-- TODO: Replace with `defaultPossessionProofGen` once cardano-crypto-class
+-- exports it from the public API.
+genLeiosPossessionProof :: Gen (PossessionProofDSIGN BLS12381MinSigDSIGN)
+genLeiosPossessionProof = do
+  seed <- arbitrarySeedOfSize (seedSizeDSIGN (Proxy @BLS12381MinSigDSIGN))
+  let sk = genKeyDSIGNWithContext @BLS12381MinSigDSIGN Nothing seed
+  pure $ createPossessionProofDSIGN minSigPoPDST sk
 
 instance Arbitrary StakePoolRelay where
   arbitrary = genericArbitraryU